Develop 1.4.0

+++ cloud-manager/src/main/java/com/webank/ai/fatecloud/system/service/facade/FederatedExchangeServiceFacade.java
@@ -22,13 +22,15 @@ import com.webank.ai.fatecloud.common.CheckSignature;
 import com.webank.ai.fatecloud.common.CommonResponse;
 import com.webank.ai.fatecloud.common.Dict;
 import com.webank.ai.fatecloud.common.Enum.ReturnCodeEnum;
+import com.webank.ai.fatecloud.common.util.ObjectUtil;
 import com.webank.ai.fatecloud.common.util.PageBean;
-import com.webank.ai.fatecloud.system.dao.entity.PartyDo;
 import com.webank.ai.fatecloud.system.dao.entity.FederatedExchangeDo;
-import com.webank.ai.fatecloud.system.dao.entity.RollSiteDo;
+import com.webank.ai.fatecloud.system.dao.entity.PartyDo;
 import com.webank.ai.fatecloud.system.pojo.dto.RollSitePageDto;
+import com.webank.ai.fatecloud.system.pojo.dto.SiteDetailDto;
 import com.webank.ai.fatecloud.system.pojo.qo.*;
 import com.webank.ai.fatecloud.system.service.impl.FederatedExchangeService;
+import com.webank.ai.fatecloud.system.service.impl.FederatedSiteManagerService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -36,9 +38,9 @@ import org.springframework.stereotype.Service;
 
 import javax.servlet.http.HttpServletRequest;
 import java.io.Serializable;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Service
 @Slf4j
@@ -46,6 +48,9 @@ public class FederatedExchangeServiceFacade implements Serializable {
 
     @Autowired
     FederatedExchangeService federatedExchangeService;
+
+    @Autowired
+    FederatedSiteManagerService federatedSiteManagerService;
 
     @Autowired
     CheckSignature checkSignature;
@@ -66,6 +71,22 @@ public class FederatedExchangeServiceFacade implements Serializable {
         if (i == 3) {
             return new CommonResponse<>(ReturnCodeEnum.ROLLSITE_NETWORK_REPEAT);
         }
+
+        for (RollSiteAddBean rollSiteAddBean : exchangeAddQo.getRollSiteAddBeanList()) {
+            if (federatedExchangeService.queryExchange(new ExchangeQueryQo(rollSiteAddBean.getNetworkAccess())) == null) {
+                return new CommonResponse<>(ReturnCodeEnum.ROLLSITE_GRPC_ERROR, rollSiteAddBean.getNetworkAccess());
+            }
+
+            // remove admin exchange party, check party ambiguity
+            List<PartyAddBean> partyAddBeanList = rollSiteAddBean.getPartyAddBeanList();
+            partyAddBeanList.removeIf(partyAddBean -> federatedExchangeService.checkExchangePartyId(partyAddBean.getPartyId()));
+            for (PartyAddBean partyAddBean : partyAddBeanList) {
+                if (federatedExchangeService.checkPartyExistInOtherExchange(partyAddBean.getPartyId(), null)) {
+                    return new CommonResponse<>(ReturnCodeEnum.SITE_PARTY_EXCHANGE_REPEAT_ERROR, partyAddBean.getPartyId());
+                }
+            }
+        }
+
         FederatedExchangeDo federatedExchangeDo = federatedExchangeService.addExchange(exchangeAddQo);
 
         if (federatedExchangeDo == null) {
@@ -82,7 +103,7 @@ public class FederatedExchangeServiceFacade implements Serializable {
             return 1;
         }
 
-        if (StringUtils.isBlank(exchangeAddQo.getVip())) {
+        if (!ObjectUtil.matchNetworkAddress(exchangeAddQo.getVipEntrance())) {
             return 1;
         }
 
@@ -97,9 +118,9 @@ public class FederatedExchangeServiceFacade implements Serializable {
 
             //check roll site ip
             String networkAccess = rollSiteAddBean.getNetworkAccess();
-//            if (!networkAccess.matches(ipAndPortRegex)) {
-//                return 1;
-//            }
+            if (!ObjectUtil.matchNetworkAddress(networkAccess) || !ObjectUtil.matchNetworkAddress(rollSiteAddBean.getNetworkAccessExit())) {
+                return 1;
+            }
 
             //check roll site is managed or not
             if (federatedExchangeService.findRollSite(networkAccess)) {
@@ -121,9 +142,9 @@ public class FederatedExchangeServiceFacade implements Serializable {
             return 3;
         }
 
-//        if (!exchangeAddQo.getVip().matches(ipAndPortRegex)) {
-//            return 1;
-//        }
+        //if (!exchangeAddQo.getVipEntrance().matches("[\\d]{1,3}(\\.[\\d]{1,3}){3}:[0-9]{1,5}")) {
+        //    return 1;
+        //}
 
         return 0;
 
@@ -138,6 +159,12 @@ public class FederatedExchangeServiceFacade implements Serializable {
     }
 
     public CommonResponse deleteExchange(ExchangeDeleteQo exchangeDeleteQo) {
+        // check if there are any using sites under the switch
+        boolean result = federatedExchangeService.checkSiteExistByExchange(exchangeDeleteQo.getExchangeId());
+        if (result) {
+            return new CommonResponse<>(ReturnCodeEnum.SITE_PARTY_EXIST_ERROR);
+        }
+
         federatedExchangeService.deleteExchange(exchangeDeleteQo);
         return new CommonResponse<>(ReturnCodeEnum.SUCCESS);
 
@@ -153,10 +180,16 @@ public class FederatedExchangeServiceFacade implements Serializable {
         HashSet<String> partyIds = new HashSet<>();
 
         for (PartyAddBean partyAddBean : partyAddBeanList) {
+            if (federatedExchangeService.checkExchangePartyId(partyAddBean.getPartyId())) {
+                partyIds.add(partyAddBean.getPartyId());
+                continue;
+            }
+
             String networkAccess = partyAddBean.getNetworkAccess();
-//            if (!networkAccess.matches(ipAndPortRegex)) {
-//                return false;
-//            }
+            if (!ObjectUtil.matchNetworkAddressNew(networkAccess) && !ObjectUtil.matchNetworkAddress(networkAccess)) {
+                return false;
+            }
+
             //check secure status
             Integer secureStatus = partyAddBean.getSecureStatus();
             if (secureStatus == null || (secureStatus != 1 && secureStatus != 2)) {
@@ -179,9 +212,7 @@ public class FederatedExchangeServiceFacade implements Serializable {
                 Integer status = partyAddBean.getStatus();
 
                 HashSet<Integer> integers = new HashSet<>();
-                integers.add(1);
-                integers.add(2);
-                integers.add(3);
+                Collections.addAll(integers, 1, 2, 3);
                 if (!integers.contains(status)) {
                     return false;
                 }
@@ -189,11 +220,7 @@ public class FederatedExchangeServiceFacade implements Serializable {
             }
         }
 
-        if (partyIds.size() != partyAddBeanList.size()) {
-            return false;
-        }
-
-        return true;
+        return partyIds.size() == partyAddBeanList.size();
     }
 
 
@@ -218,7 +245,7 @@ public class FederatedExchangeServiceFacade implements Serializable {
     public CommonResponse<List<PartyDo>> queryExchange(ExchangeQueryQo exchangeQueryQo) {
 
         String networkAccess = exchangeQueryQo.getNetworkAccess();
-        if (StringUtils.isBlank(networkAccess)) {
+        if (!ObjectUtil.matchNetworkAddress(networkAccess)) {
             return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
 
         }
@@ -244,9 +271,22 @@ public class FederatedExchangeServiceFacade implements Serializable {
         if (federatedExchangeService.findRollSite(rollSieAddQo.getNetworkAccess())) {
             return new CommonResponse<>(ReturnCodeEnum.ROLLSITE_EXIST_ERROR);
         }
+        if (federatedExchangeService.queryExchange(new ExchangeQueryQo(rollSieAddQo.getNetworkAccess())) == null) {
+            return new CommonResponse<>(ReturnCodeEnum.ROLLSITE_GRPC_ERROR, rollSieAddQo.getNetworkAccess());
+        }
+
         //check party
-        if (!this.checkPartyNetwork(rollSieAddQo.getPartyAddBeanList(), true)) {
+        List<PartyAddBean> partyAddBeanList = rollSieAddQo.getPartyAddBeanList();
+        if (!this.checkPartyNetwork(partyAddBeanList, true) || !ObjectUtil.matchNetworkAddress(rollSieAddQo.getNetworkAccessExit())) {
             return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
+        }
+
+        // remove admin exchange party, check party ambiguity
+        partyAddBeanList.removeIf(partyAddBean -> federatedExchangeService.checkExchangePartyId(partyAddBean.getPartyId()));
+        for (PartyAddBean partyAddBean : partyAddBeanList) {
+            if (federatedExchangeService.checkPartyExistInOtherExchange(partyAddBean.getPartyId(), rollSieAddQo.getExchangeId())) {
+                return new CommonResponse<>(ReturnCodeEnum.SITE_PARTY_EXCHANGE_REPEAT_ERROR);
+            }
         }
 
         federatedExchangeService.addRollSite(rollSieAddQo);
@@ -266,6 +306,8 @@ public class FederatedExchangeServiceFacade implements Serializable {
             return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
         }
 
+        // exclude exchange admin party
+        rollSiteUpdateQo.getPartyAddBeanList().removeIf(partyAddBean -> federatedExchangeService.checkExchangePartyId(partyAddBean.getPartyId()));
         federatedExchangeService.updateRollSite(rollSiteUpdateQo);
         return new CommonResponse<>(ReturnCodeEnum.SUCCESS);
     }
@@ -276,6 +318,11 @@ public class FederatedExchangeServiceFacade implements Serializable {
             return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
 
         }
+
+        /*if (federatedExchangeService.checkSiteExistByRollSite(rollSiteId)) {
+            return new CommonResponse<>(ReturnCodeEnum.SITE_PARTY_EXIST_ERROR);
+        }*/
+
         federatedExchangeService.deleteRollSite(rollSiteDeleteQo);
         return new CommonResponse<>(ReturnCodeEnum.SUCCESS);
     }
@@ -311,5 +358,65 @@ public class FederatedExchangeServiceFacade implements Serializable {
 
         return new CommonResponse<>(ReturnCodeEnum.SUCCESS, rollSitePageDtoPageBean);
 
+    }
+
+    // v1.4
+    public CommonResponse<List<PartyDo>> finPartyPage(PartyQueryQo partyQueryQo) {
+        if (ObjectUtil.isEmpty(partyQueryQo.getRollSiteId())) {
+            return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
+        }
+
+        PageBean<PartyDo> partyDoPageBean = federatedExchangeService.findPartyPage(partyQueryQo);
+        return new CommonResponse<>(ReturnCodeEnum.SUCCESS, partyDoPageBean.getList());
+    }
+
+    // v1.4
+    public CommonResponse<Void> updateIpManagerPartyInfo(PartyUpdateQo partyUpdateQo) {
+        if (ObjectUtil.isEmpty(partyUpdateQo.getId(), partyUpdateQo.getPartyId(), partyUpdateQo.getExchangeId(),
+                partyUpdateQo.getPollingStatus(), partyUpdateQo.getSecureStatus()) ||
+                !ObjectUtil.matchNetworkAddressNew(partyUpdateQo.getNetworkAccessEntrances())) {
+            return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
+        }
+
+        try {
+            federatedExchangeService.updateIpManagerPartyInfo(partyUpdateQo);
+        } catch (Exception e) {
+            log.error(e.getMessage(), e);
+            return new CommonResponse<>(ReturnCodeEnum.SITE_PARTY_UPDATE_ERROR);
+        }
+        return new CommonResponse<>(ReturnCodeEnum.SUCCESS);
+    }
+
+    // v1.4
+    public CommonResponse<List<FederatedExchangeDo>> findAllExchange() {
+        List<FederatedExchangeDo> exchangeDoList = federatedExchangeService.findAllExchange();
+        return new CommonResponse<>(ReturnCodeEnum.SUCCESS, exchangeDoList);
+    }
+
+    public CommonResponse<Boolean> checkPartyExist(PartyQueryQo partyQueryQo) {
+        if (ObjectUtil.isEmpty(partyQueryQo.getPartyId())) {
+            return new CommonResponse<>(ReturnCodeEnum.PARAMETERS_ERROR);
+        }
+
+        if (federatedExchangeService.checkExchangePartyId(partyQueryQo.getPartyId())) {
+            return new CommonResponse<>(ReturnCodeEnum.SUCCESS, Boolean.TRUE);
+        }
+
+        long parseInt;
+        try {
+            parseInt = Long.parseLong(partyQueryQo.getPartyId());
+        } catch (Exception e) {
+            return new CommonResponse<>(ReturnCodeEnum.SUCCESS, Boolean.FALSE);
+        }
+
+        if (partyQueryQo.getRollSiteId() != null) {
+            List<PartyDo> partyByRollSiteId = federatedExchangeService.findPartyByRollSiteId(partyQueryQo.getRollSiteId());
+            if (partyByRollSiteId.stream().anyMatch(partyDo -> partyQueryQo.getPartyId().equals(partyDo.getPartyId()))) {
+                return new CommonResponse<>(ReturnCodeEnum.SUCCESS, Boolean.TRUE);
+            }
+        }
+
+        SiteDetailDto siteByPartyId = federatedSiteManagerService.findSiteByPartyId(parseInt, "", 2);
+        return new CommonResponse<>(ReturnCodeEnum.SUCCESS, siteByPartyId != null);
     }
 }
